Remove hard dependency on nvim-treesitter plugin
Markview has a hard dependency on the nvim-treesitter plugin (ie. the
plugin won't load without it). This is unnecessary because the true
dependency is on the markdown and markdown_inline treesitter parsers,
not on the nvim-treesitter plugin itself. The check should be changed to
use the vim.treesitter functions instead to check for dependencies and
not on the nvim-treesitter plugin. This allows for the use of other
treesitter parser installers such as that found on rocks.nvim.