fix(renderer_markdown): apply `scope_hl` when concealing is disabled
When using a `scope_hl` in `markdown_inline.checkboxes`, it was
previously ignored when concealing was also disabled with
`conceal_on_checkboxes = false` in `markdown.list_items`.

The concealing only applies to the list bullets/numbers, so we can still
apply the `scope_hl` to the text of the list item.